+++ gestiondestock/src/main/java/com/tehjul/gestiondestock/services/impl/CommandeFournisseurServiceImpl.java
@@ -236,6 +236,10 @@ public class CommandeFournisseurServiceImpl implements CommandeFournisseurServic
             log.error("Commande fournisseur ID is NULL");
             return;
         }
+        List<LigneCommandeFournisseur> ligneCommandeFournisseurs= ligneCommandeFournisseurRepository.findAllByCommandeFournisseurId(id);
+        if (!ligneCommandeFournisseurs.isEmpty()) {
+            throw new InvalidOperationException("Impossible de supprimer une commande fournisseur déjà utilisée", ErrorCodes.COMMANDE_FOURNISSEUR_ALREADY_IN_USE);
+        }
         commandeFournisseurRepository.deleteById(id);
     }
